Evmcontract

查看合并请求 Higgs.Trust/trust!140

+++ slave/src/main/java/com/higgs/trust/slave/core/repository/PackageRepository.java
@@ -1,19 +1,27 @@
 package com.higgs.trust.slave.core.repository;
 
-import com.google.common.collect.Lists;
+import com.higgs.trust.common.constant.Constant;
+import com.higgs.trust.common.dao.RocksUtils;
+import com.higgs.trust.common.utils.ThreadLocalUtils;
+import com.higgs.trust.slave.common.config.InitConfig;
 import com.higgs.trust.slave.common.enums.SlaveErrorEnum;
 import com.higgs.trust.slave.common.exception.SlaveException;
-import com.higgs.trust.slave.dao.pack.PackageDao;
+import com.higgs.trust.slave.core.repository.config.SystemPropertyRepository;
+import com.higgs.trust.slave.dao.mysql.pack.PackageDao;
 import com.higgs.trust.slave.dao.po.pack.PackagePO;
+import com.higgs.trust.slave.dao.rocks.pack.PackRocksDao;
+import com.higgs.trust.slave.dao.rocks.pack.PackStatusRocksDao;
 import com.higgs.trust.slave.model.bo.Package;
+import com.higgs.trust.slave.model.bo.config.SystemProperty;
 import com.higgs.trust.slave.model.enums.biz.PackageStatusEnum;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.rocksdb.ReadOptions;
+import org.rocksdb.Transaction;
+import org.rocksdb.WriteOptions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.TransactionStatus;
-import org.springframework.transaction.support.TransactionCallbackWithoutResult;
-import org.springframework.transaction.support.TransactionTemplate;
 
 import java.util.List;
 import java.util.Set;
@@ -23,11 +31,18 @@ import java.util.Set;
  * @author: pengdi
  **/
 @Repository @Slf4j public class PackageRepository {
-    @Autowired private TransactionTemplate txRequired;
 
     @Autowired PackageDao packageDao;
 
+    @Autowired PackRocksDao packRocksDao;
+
+    @Autowired PackStatusRocksDao packStatusRocksDao;
+
+    @Autowired SystemPropertyRepository systemPropertyRepository;
+
     @Autowired PendingTxRepository pendingTxRepository;
+
+    @Autowired InitConfig initConfig;
 
     /**
      * new package from repository
@@ -40,13 +55,14 @@ import java.util.Set;
             return;
         }
 
-        //TODO rocks db
         PackagePO packagePO = convertPackToPackagePO(pack);
-        packageDao.add(packagePO);
-    }
-
-    public PackagePO queryByHeight(Long height) {
-        return packageDao.queryByHeight(height);
+        if (initConfig.isUseMySQL()) {
+            packageDao.add(packagePO);
+        } else {
+            packRocksDao.save(packagePO);
+            //to store package status
+            packStatusRocksDao.save(packagePO.getHeight(), pack.getStatus().getIndex());
+        }
     }
 
     /**
@@ -57,15 +73,15 @@ import java.util.Set;
      * @param to
      */
     public void updateStatus(Long height, PackageStatusEnum from, PackageStatusEnum to) {
-        txRequired.execute(new TransactionCallbackWithoutResult() {
-            @Override protected void doInTransactionWithoutResult(TransactionStatus status) {
-                int r = packageDao.updateStatus(height, from.getCode(), to.getCode());
-                if (r != 1) {
-                    log.error("package.updateStatus is fail from:{} to:{}",from,to);
-                    throw new SlaveException(SlaveErrorEnum.SLAVE_PACKAGE_UPDATE_STATUS_ERROR);
-                }
+        if (initConfig.isUseMySQL()) {
+            int r = packageDao.updateStatus(height, from.getCode(), to.getCode());
+            if (r != 1) {
+                log.error("[package.updateStatus] has error");
+                throw new SlaveException(SlaveErrorEnum.SLAVE_PACKAGE_UPDATE_STATUS_ERROR);
             }
-        });
+        } else {
+            packStatusRocksDao.update(height, from.getIndex(), to.getIndex());
+        }
     }
 
     /**
@@ -90,12 +106,19 @@ import java.util.Set;
      * @return
      */
     public Package load(Long height) {
-        PackagePO packagePO = packageDao.queryByHeight(height);
+        PackagePO packagePO;
 
-        if (null == packagePO) {
-            return null;
+        if (initConfig.isUseMySQL()) {
+            packagePO = packageDao.queryByHeight(height);
+            if (null != packagePO) {
+                packagePO.setSignedTxList(pendingTxRepository.getTransactionsByHeight(packagePO.getHeight()));
+            }
+        } else {
+            packagePO = packRocksDao.get(String.valueOf(height));
+            if (null != packagePO) {
+                packagePO.setStatus(packStatusRocksDao.getStatusByHeight(height));
+            }
         }
-
         return convertPackagePOToPackage(packagePO);
     }
 
@@ -106,12 +129,12 @@ import java.util.Set;
      * @return
      */
     public List<Long> loadHeightList(Long height) {
-        List<Long> heights = packageDao.queryHeightListByHeight(height);
-
-        if (CollectionUtils.isEmpty(heights)) {
-            return null;
+        List<Long> heights;
+        if (initConfig.isUseMySQL()) {
+            heights = packageDao.queryHeightListByHeight(height);
+        } else {
+            heights = packStatusRocksDao.queryHeightListByHeight(height + 1);
         }
-
         return heights;
     }
 
@@ -122,13 +145,16 @@ import java.util.Set;
      * @return
      */
     public Package loadAndLock(Long height) {
+        PackagePO packagePO;
 
-        PackagePO packagePO = packageDao.queryByHeightForUpdate(height);
-
-        if (null == packagePO) {
-            return null;
+        if (initConfig.isUseMySQL()) {
+            packagePO = packageDao.queryByHeightForUpdate(height);
+            if (null != packagePO) {
+                packagePO.setSignedTxList(pendingTxRepository.getTransactionsByHeight(packagePO.getHeight()));
+            }
+        } else {
+            packagePO = packRocksDao.get(String.valueOf(height));
         }
-
         return convertPackagePOToPackage(packagePO);
     }
 
@@ -138,68 +164,27 @@ import java.util.Set;
      * @return
      */
     public Long getMaxHeight() {
-        return packageDao.getMaxHeight();
-    }
-
-    /**
-     * get package list
-     *
-     * @return
-     */
-    public List<Long> getHeightListByStatus(String status) {
-        List<Long> heightList = packageDao.queryHeightListByStatus(status);
-
-        if (CollectionUtils.isEmpty(heightList)) {
-            return null;
+        if (initConfig.isUseMySQL()) {
+            return packageDao.getMaxHeight();
+        } else {
+            SystemProperty bo = systemPropertyRepository.queryByKey(Constant.MAX_PACK_HEIGHT);
+            return bo != null && !StringUtils.isEmpty(bo.getValue()) ? Long.parseLong(bo.getValue()) : null;
         }
-
-        return heightList;
     }
 
-    /**
-     * get package list
-     *
-     * @return
-     */
-    public List<Long> getHeightsByStatusAndLimit(String status, int limit) {
-        List<Long> heightList = packageDao.queryHeightsByStatusAndLimit(status, limit);
-
-        if (CollectionUtils.isEmpty(heightList)) {
-            return null;
-        }
-
-        return heightList;
-    }
-
-    /**
-     * get min height with package status
-     *
-     * @param statusSet
-     * @return
-     */
-    public Long getMinHeight(Set<String> statusSet, Long maxBlockHeight) {
-        return packageDao.getMinHeightWithStatus(statusSet, maxBlockHeight);
-    }
-
-    /**
-     * get min height with package status and start height
-     *
-     * @param startHeight
-     * @param statusSet
-     * @return
-     */
-    public Long getMinHeight(long startHeight, Set<String> statusSet) {
-        return packageDao.getMinHeightWithHeightAndStatus(startHeight, statusSet);
-    }
-
-    /**
+     /**
      * get count with package status
      *
      * @param statusSet
      * @return
      */
     public long count(Set<String> statusSet, Long maxBlockHeight) {
-        return packageDao.countWithStatus(statusSet, maxBlockHeight);
+        if (initConfig.isUseMySQL()) {
+            return packageDao.countWithStatus(statusSet, maxBlockHeight);
+        } else {
+            SystemProperty bo = systemPropertyRepository.queryByKey(Constant.MAX_PACK_HEIGHT);
+            return bo != null && !StringUtils.isEmpty(bo.getValue()) ? Long.parseLong(bo.getValue()) - maxBlockHeight : 0;
+        }
     }
 
     /**
@@ -213,6 +198,7 @@ import java.util.Set;
         packagePO.setHeight(pack.getHeight());
         packagePO.setPackageTime(pack.getPackageTime());
         packagePO.setStatus(pack.getStatus().getCode());
+        packagePO.setSignedTxList(pack.getSignedTxList());
         return packagePO;
     }
 
@@ -223,32 +209,16 @@ import java.util.Set;
      * @return
      */
     private Package convertPackagePOToPackage(PackagePO packagePO) {
+        if (null == packagePO) {
+            return null;
+        }
         Package packageBO = new Package();
         packageBO.setHeight(packagePO.getHeight());
         packageBO.setPackageTime(packagePO.getPackageTime());
         packageBO.setStatus(PackageStatusEnum.getByCode(packagePO.getStatus()));
 
-        packageBO.setSignedTxList(pendingTxRepository.getTransactionsByHeight(packagePO.getHeight()));
+        packageBO.setSignedTxList(packagePO.getSignedTxList());
         return packageBO;
-    }
-
-    /**
-     * convert packagePO to package
-     *
-     * @param packagePOs
-     * @return
-     */
-    private List<Package> convertPackagePOsToPackages(List<PackagePO> packagePOs) {
-        List<Package> packages = Lists.newLinkedList();
-        for (PackagePO packagePO : packagePOs) {
-            Package pack = new Package();
-            pack.setHeight(packagePO.getHeight());
-            pack.setPackageTime(packagePO.getPackageTime());
-            pack.setStatus(PackageStatusEnum.getByCode(packagePO.getStatus()));
-            pack.setSignedTxList(pendingTxRepository.getTransactionsByHeight(packagePO.getHeight()));
-            packages.add(pack);
-        }
-        return packages;
     }
 
     /**
@@ -259,13 +229,22 @@ import java.util.Set;
      * @return
      */
     public boolean isPackageStatus(Long height, PackageStatusEnum packageStatusEnum) {
-        PackagePO packagePO = queryByHeight(height);
+        PackagePO packagePO = null;
+        if (initConfig.isUseMySQL()) {
+            packagePO = packageDao.queryByHeight(height);
+        } else {
+            packagePO = packRocksDao.get(String.valueOf(height));
+            if (null != packagePO) {
+                packagePO.setStatus(packStatusRocksDao.getStatusByHeight(height));
+            }
+        }
         if (packagePO == null) {
             log.warn("[isPackageStatus] package is null height:{}", height);
             return false;
         }
-        log.debug("package of DB status:{},blockHeight:{}", packagePO.getStatus(), height);
-        return PackageStatusEnum.getByCode(packagePO.getStatus()) == packageStatusEnum;
+        PackageStatusEnum status = PackageStatusEnum.getByCode(packagePO.getStatus());
+        log.debug("package of DB status:{}, blockHeight:{}", status, height);
+        return status == packageStatusEnum;
     }
 
     /**
@@ -275,7 +254,10 @@ import java.util.Set;
      * @return
      */
     public Long getMaxHeightByStatus(PackageStatusEnum status) {
-        return packageDao.getMaxHeightByStatus(status.getCode());
+        if (initConfig.isUseMySQL()) {
+            return packageDao.getMaxHeightByStatus(status.getCode());
+        }
+        return packStatusRocksDao.getMaxHeightByStatus(status.getIndex());
     }
 
     /**
@@ -285,7 +267,10 @@ import java.util.Set;
      * @return
      */
     public Long getMinHeightByStatus(PackageStatusEnum status) {
-        return packageDao.getMinHeightByStatus(status.getCode());
+        if (initConfig.isUseMySQL()) {
+            return packageDao.getMinHeightByStatus(status.getCode());
+        }
+        return packStatusRocksDao.getMinHeightByStatus(status.getIndex());
     }
 
     /**
@@ -295,6 +280,52 @@ import java.util.Set;
      * @return
      */
     public int deleteLessThanHeightAndStatus(Long height,PackageStatusEnum status) {
-        return packageDao.deleteLessThanHeightAndStatus(height,status.getCode());
+        if (initConfig.isUseMySQL()) {
+            return packageDao.deleteLessThanHeightAndStatus(height, status.getCode());
+        }
+
+        List<Long> heights = packStatusRocksDao.queryByPrefix(status.getIndex());
+        int count = 0;
+        Transaction tx = RocksUtils.beginTransaction(new WriteOptions());
+        ThreadLocalUtils.putRocksTx(tx);
+        try {
+            for (Long h : heights) {
+                if (height.compareTo(h) >= 0) {
+                    deletePendingTx(height);
+                    packRocksDao.batchDelete(h);
+                    packStatusRocksDao.batchDelete(h, status.getIndex());
+                    count++;
+                }
+            }
+            RocksUtils.txCommit(tx);
+        } finally {
+            ThreadLocalUtils.clearRocksTx();
+        }
+        return count;
+    }
+
+    public Long getForUpdate(Transaction tx, ReadOptions readOptions, String key, boolean exclusive) {
+        return packStatusRocksDao.getForUpdate(tx, readOptions, key, exclusive);
+    }
+
+    private void deletePendingTx(Long height) {
+        PackagePO po = packRocksDao.get(String.valueOf(height));
+        if (po == null || CollectionUtils.isEmpty(po.getSignedTxList())) {
+            return;
+        }
+
+        pendingTxRepository.batchDelete(po.getSignedTxList());
+    }
+
+    /**
+     * get height list by status
+     * @param status
+     * @return
+     */
+    public List<Long> getBlockHeightsByStatus(PackageStatusEnum status) {
+        if (initConfig.isUseMySQL()) {
+            return packageDao.getBlockHeightsByStatus(status.getCode());
+        }
+        return packStatusRocksDao.getBlockHeightsByStatus(status.getIndex());
     }
 }
